Be more tolerant of graphite service restarts
Network failures may cause the graphite client to restart which causes the data to slightly out of date until the next data push cycle. Update the graphite data age threshold in the check-data.rb script (-a) to 90s to better tolerate these blips and not alert immediately. Can't use spaces in the parameters due to the way batch and then Ruby handle the tokenization of the parameters.